Create disbursement logic

+++ db/migrate/20221123113646_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
-      t.decimal :amount, null: false
+      t.float :amount, null: false
       t.timestamp :completed_at
       t.references :merchant, foreign_key: true
       t.references :shopper, foreign_key: true
