+++ db/migrate/20221123114332_create_disbursements.rb
@@ -1,7 +1,7 @@
 class CreateDisbursements < ActiveRecord::Migration[7.0]
   def change
     create_table :disbursements do |t|
-      t.decimal :amount, default: 0.0
+      t.float :amount, default: 0.0
       t.references :merchant , foreign_key: true
       t.integer :week
       t.integer :year
